@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-from typing import Any, Dict, List
+from typing import Any
 
 from _pytest.fixtures import FixtureDef
 
-FixtureDict = Dict[str, List[FixtureDef[Any]]]
+FixtureDict = dict[str, list[FixtureDef[Any]]]
